+++ tests/test_prg_engine_control_flow_support.h
@@ -131,6 +131,7 @@ void test_total_command_supports_currency_and_integer_fields();
 void test_total_command_for_sql_result_cursors();
 void test_count_for_expression_closing_target_cursor_fails_catchably();
 void test_sum_value_expression_closing_target_cursor_fails_catchably();
+void test_sum_multiple_targets_preserves_earlier_targets_when_later_expression_closes_cursor();
 void test_total_for_expression_closing_target_cursor_fails_catchably();
 void test_select_query_into_array_commands();
 void test_create_table_free_dynamic_target();